:feelsgood: Adding main.js for configuration. Loading temp-full-config.json for all configurations
There is now a chartingApp object in the global namespace that will
store the configuration object so that it can be read from any module or
any other piece of code. Not sure if this is the best way, but if we
keep this one object slim, it should be an ok way to store the
configuration for modules.